requirements.txt
starkbank-ecdsa==0.1.4
python-telegram-bot==13.0
websocket-client==0.59.0